feat(auth-node): setup base auth-node package (#main)
This commit adds the base setup for the auth-node package. It includes new files such as README, jest and rollup configurations, package.json, and several TypeScript files defining the base client and types for the package.

It also includes an integration smoke test, and extends the types from auth-js package. The base client provides a get context function while the index.ts exports the necessary classes and types.