feat: add URL handling and update logout functionality (#main)
In the client.ts file of both browser and client packages, a new method 'openUrlWithFallback' has been added which assigns the given url to the window's location. In the client.ts file of the client package, the 'logout' method has been updated to handle the url opening logic through the 'openUrlWithFallback' method. The 'logout' method also now supports 'LogoutUrlOptions'. The 'openUrlWithFallback' method in the 'client.ts' file now throws an error if the 'openUrl' method is not implemented. The '_url' method has been renamed to '_buildUrl' and a new method '_buildLogoutUrl' has been added for creating the logout url. The 'server-actions.ts' file in the next package no longer uses the 'server' import. In the 'client.logout.unit.ts' file in the client package, a new test has been added for the 'logout'